--- a/robot-server/tests/labware_offsets/test_store.py
+++ b/robot-server/tests/labware_offsets/test_store.py
@@ -56,13 +56,15 @@ def subject(
 def test_empty_search(subject: LabwareOffsetStore) -> None:
     """Searching with no filters should return no results."""
     subject.add(
-        IncomingStoredLabwareOffset(
-            id="id",
-            createdAt=datetime.now(timezone.utc),
-            definitionUri="namespace/load_name/1",
-            locationSequence="anyLocation",
-            vector=LabwareOffsetVector(x=1, y=2, z=3),
-        )
+        [
+            IncomingStoredLabwareOffset(
+                id="id",
+                createdAt=datetime.now(timezone.utc),
+                definitionUri="namespace/load_name/1",
+                locationSequence="anyLocation",
+                vector=LabwareOffsetVector(x=1, y=2, z=3),
+            )
+        ]
     )
     assert subject.search([]) == []
 
@@ -79,13 +81,15 @@ def test_search_most_recent_only(subject: LabwareOffsetStore) -> None:
     ]
     for id, definition_uri in ids_and_definition_uris:
         subject.add(
-            IncomingStoredLabwareOffset(
-                id=id,
-                definitionUri=definition_uri,
-                createdAt=datetime.now(timezone.utc),
-                locationSequence="anyLocation",
-                vector=LabwareOffsetVector(x=1, y=2, z=3),
-            )
+            [
+                IncomingStoredLabwareOffset(
+                    id=id,
+                    definitionUri=definition_uri,
+                    createdAt=datetime.now(timezone.utc),
+                    locationSequence="anyLocation",
+                    vector=LabwareOffsetVector(x=1, y=2, z=3),
+                )
+            ]
         )
 
     results = subject.search([SearchFilter(mostRecentOnly=True)])
@@ -263,7 +267,7 @@ def test_filter_fields(
         ),
     }
     for offset in offsets.values():
-        subject.add(offset)
+        subject.add([offset])
     results = subject.search(
         [
             SearchFilter(
@@ -323,8 +327,7 @@ def test_filter_field_combinations(subject: LabwareOffsetStore) -> None:
         for offset in labware_offsets
     ]
 
-    for labware_offset in labware_offsets:
-        subject.add(labware_offset)
+    subject.add(labware_offsets)
 
     # Filter accepting any value for each field (i.e. a return-everything filter):
     result = subject.search([SearchFilter()])
@@ -368,13 +371,15 @@ def test_filter_combinations(subject: LabwareOffsetStore) -> None:
     ]
     for id, definition_uri in ids_and_definition_uris:
         subject.add(
-            IncomingStoredLabwareOffset(
-                id=id,
-                createdAt=datetime.now(timezone.utc),
-                definitionUri=definition_uri,
-                locationSequence=ANY_LOCATION,
-                vector=LabwareOffsetVector(x=1, y=2, z=3),
-            )
+            [
+                IncomingStoredLabwareOffset(
+                    id=id,
+                    createdAt=datetime.now(timezone.utc),
+                    definitionUri=definition_uri,
+                    locationSequence=ANY_LOCATION,
+                    vector=LabwareOffsetVector(x=1, y=2, z=3),
+                )
+            ]
         )
 
     # Multiple filters should be OR'd together.
@@ -429,9 +434,7 @@ def test_delete(subject: LabwareOffsetStore) -> None:
     with pytest.raises(LabwareOffsetNotFoundError):
         subject.delete("b")
 
-    subject.add(a)
-    subject.add(b)
-    subject.add(c)
+    subject.add([a, b, c])
 
     assert subject.delete(b.id) == out_b
     assert subject.get_all() == [out_a, out_c]
@@ -473,7 +476,7 @@ def test_handle_unknown(
         ],
         vector=incoming_valid.vector,
     )
-    subject.add(incoming_valid)
+    subject.add([incoming_valid])
     with sql_engine.begin() as transaction:
         transaction.execute(
             sqlalchemy.insert(labware_offset_location_sequence_components_table).values(
@@ -496,13 +499,15 @@ def test_notifications(
     """It should publish notifications any time the set of labware offsets changes."""
     decoy.verify(mock_labware_offsets_publisher.publish_labware_offsets(), times=0)
     subject.add(
-        IncomingStoredLabwareOffset(
-            id="id",
-            createdAt=datetime.now(timezone.utc),
-            definitionUri="definitionUri",
-            locationSequence=ANY_LOCATION,
-            vector=LabwareOffsetVector(x=1, y=2, z=3),
-        )
+        [
+            IncomingStoredLabwareOffset(
+                id="id",
+                createdAt=datetime.now(timezone.utc),
+                definitionUri="definitionUri",
+                locationSequence=ANY_LOCATION,
+                vector=LabwareOffsetVector(x=1, y=2, z=3),
+            )
+        ]
     )
     decoy.verify(mock_labware_offsets_publisher.publish_labware_offsets(), times=1)
     subject.delete("id")