## Overview

An HTTP client can upload multiple labware offsets to the robot server in a single `POST /labwareOffsets` request. Formerly, each offset was running in its own SQL transaction, which was slow. This wraps all of the insertions in a single transaction, which is much faster.

Inserting 1000 offsets on an OT-2 goes from ~210s (~210ms/offset) to ~22s (~22ms/offset).

## Test Plan and Hands on Testing

We're covered by existing automated tests, but I also tried this manually on an OT-2.

## Review requests

None in particular.

## Risk assessment

Low. Changes are pretty trivial.